+++ app/Resources/static-src/app/js/admin-v2/main/index.js
@@ -1,0 +1,2 @@
+import 'codeages-design';
+import './menu-mark-new';