+++ web/bundles/translations/en.js
@@ -1310,4 +1310,12 @@ t.add("admin.mobile_manage.discovery_upgrade_madal.notice_successed", "Upgrade s
 t.add("admin.mobile_manage.discovery_upgrade_madal.notice_failed", "Upgrade failed", "js", "en");
 t.add("admin.mobile_manage.discovery_upgrade_madal.notice_upgraded", "Cannot repeat upgrade", "js", "en");
 t.add("course.plan_task.lessons_publish_status", "Published Lesson Number:%publishedNum%<br\/>Unpublish Lesson Number:%unpublishedNum%", "js", "en");
+t.add("admin_v2.user.cloude_sms_enable_hint", "please go to\u3010Admin\u3011-\u3010App Cente\u3011-\u3010Cloud service\u3011-\u3010Cloud Sms\u3011open cloud sms first.", "js", "en");
+t.add("admin_v2.homepage.quick_entrance.hint", "Set up to 7 quick entries", "js", "en");
+t.add("admin_v2.homepage.mini_program.title", "Scan code to open the miniprogram", "js", "en");
+t.add("admin.switch_old_version.permission_error", "Please contact super administrator or administrator operation", "js", "en");
+t.add("admin.switch_old_version.title", "Toggle backstage", "js", "en");
+t.add("admin.switch_old_version.confirm_message", "After switching to the old background, all administrators will switch to the old background to confirm the switch?", "js", "en");
+t.add("admin.switch_new_version.confirm_message", "After upgrading to the new background, all administrators will switch to the new background, but the permissions will not change, confirm the switch?", "js", "en");
+t.add("admin.can_not_link_data", "Unable to link to data", "js", "en");
 })(Translator);