+++ web/static-dist/app/js/my/learn/my-course/index.js
@@ -1,1 +1,0 @@
-webpackJsonp(["app/js/my/learn/my-course/index"],{"0e30420a8c174cc840cc":function(c,e,n){"use strict";echo.init()}},["0e30420a8c174cc840cc"]);